Refactor: Renamed types to models
