+++ PlaylistManager/Models/ViewModelBase.cs
@@ -2,7 +2,7 @@ using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using JetBrains.Annotations;
 
-namespace PlaylistManager.Types
+namespace PlaylistManager.Models
 {
     public class ViewModelBase : INotifyPropertyChanged
     {
