+++ PlaylistManager/Utilities/PlaylistLibUtils.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using BeatSaberPlaylistsLib.Blist;
 using BeatSaberPlaylistsLib.Legacy;
 using BeatSaberPlaylistsLib.Types;
-using PlaylistManager.Types;
+using PlaylistManager.Models;
 
 namespace PlaylistManager.Utilities
 {
