+++ PlaylistManager/Utilities/ViewLocator.cs
@@ -1,7 +1,7 @@
 using System;
 using Avalonia.Controls;
 using Avalonia.Controls.Templates;
-using PlaylistManager.Types;
+using PlaylistManager.Models;
 
 namespace PlaylistManager.Utilities
 {
