+++ PlaylistManager/Views/FlowCoordinator.axaml.cs
@@ -3,7 +3,7 @@ using System.Collections.ObjectModel;
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
-using PlaylistManager.Types;
+using PlaylistManager.Models;
 
 namespace PlaylistManager.Views
 {
